# docs/openapi-week3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenGuard Kids API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/enroll": {
      "post": {
        "summary": "Enroll",
        "operationId": "enroll_api_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/device/refresh": {
      "post": {
        "summary": "Refresh",
        "operationId": "refresh_api_auth_device_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/heartbeat": {
      "post": {
        "summary": "Heartbeat",
        "operationId": "heartbeat_api_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/session": {
      "post": {
        "summary": "Update Device Session",
        "operationId": "update_device_session_api_device_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceSessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/time-status": {
      "post": {
        "summary": "Update Device Time Status",
        "operationId": "update_device_time_status_api_device_time_status_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTimeStatusInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/device/requests": {
      "post": {
        "summary": "Create Time Request",
        "operationId": "create_time_request_api_device_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTimeRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/device/event-state/ack": {
      "post": {
        "summary": "Acknowledge Erasure",
        "operationId": "acknowledge_erasure_api_device_event_state_ack_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationAck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/activity-events": {
      "post": {
        "summary": "Ingest",
        "operationId": "ingest_api_device_activity_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceSessionInput": {
        "properties": {
          "active_child_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Child Id"
          },
          "active_user": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Active User"
          },
          "active_since": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Since"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DeviceSessionInput"
      },
...
      "DeviceTimeStatusInput": {
        "properties": {
          "active_child_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Child Id"
          },
          "active_user": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Active User"
          },
          "active_since": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Since"
          },
          "policy_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Policy Version"
          },
          "used_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Used Seconds"
          },
          "remaining_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Seconds"
          },
          "quota_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quota Seconds"
          },
          "extra_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Seconds"
          },
          "active_usage": {
            "type": "boolean",
            "title": "Active Usage",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy_version"
        ],
        "title": "DeviceTimeStatusInput"
      },
      "EnrollInput": {
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z2-9]{8}$",
            "title": "Code"
          },
          "display_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Display Name"
          },
          "fingerprint": {
            "type": "string",
            "maxLength": 128,
            "minLength": 16,
            "title": "Fingerprint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "display_name",
          "fingerprint"
        ],
        "title": "EnrollInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeartbeatInput": {
        "properties": {
          "policy_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Policy Version"
          },
          "used_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Used Seconds",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "policy_version"
        ],
        "title": "HeartbeatInput"
      },
...
      "RefreshInput": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "refresh_token": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Refresh Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id",
          "refresh_token"
        ],
        "title": "RefreshInput"
      },
...
    }
  }
}